- Cambiado el formato interno del Tiempo para usar long en vez de double - Añadida base para el simulador con reloj intervalar (paquete intervalThreaded). De momento únicamente sustituye el pool de threads por N hilos, cada uno con su propia lista de tareas.

+++ PSimulator/src/es/ull/isaatc/simulation/common/Resource.java
@@ -15,7 +15,7 @@ public interface Resource extends VariableStoreSimulationObject, Describable {
 	Collection<TimeTableEntry> getTimeTableEntries();
     void addTimeTableEntry(ModelCycle cycle, Time dur, ResourceType role);
     void addTimeTableEntry(ModelCycle cycle, Time dur, ArrayList<ResourceType> roleList);
-    void addTimeTableEntry(ModelCycle cycle, double dur, ResourceType role);
-    void addTimeTableEntry(ModelCycle cycle, double dur, ArrayList<ResourceType> roleList);
+    void addTimeTableEntry(ModelCycle cycle, long dur, ResourceType role);
+    void addTimeTableEntry(ModelCycle cycle, long dur, ArrayList<ResourceType> roleList);
 	
 }
